accumulator/stump: write roots len as 64-representation always
On 32-bit platform usize would be written as 32-bit little endian, and
fail to be read during deserialization (was expecting 64-bit).

+++ src/accumulator/stump.rs
@@ -107,7 +107,7 @@ impl Stump {
     pub fn serialize<Dst: Write>(&self, mut writer: &mut Dst) -> std::io::Result<usize> {
         let mut len = 8;
         writer.write_all(&self.leaves.to_le_bytes())?;
-        writer.write_all(&self.roots.len().to_le_bytes())?;
+        writer.write_all(&(self.roots.len() as u64).to_le_bytes())?;
         for root in self.roots.iter() {
             len += 32;
             root.write(&mut writer)?;
